corrected num declarations

+++ Average.java
@@ -10,10 +10,10 @@ public class Average
     double num1, num2;
 
     System.out.print( "First temperature? " );
-    num1 = keyboard.nextInt();
+    num1 = keyboard.nextDouble();
 
     System.out.print( "Second temperature? " );
-    num2 = keyboard.nextInt();
+    num2 = keyboard.nextDouble();
 
     System.out.println("The avg value is : " + (num1+num2)/2);
     keyboard.close();
